+++ libs/ajax-form/src/lib/ajax-form/ajax-form.component.ts
@@ -72,7 +72,6 @@ export class AjaxFormComponent implements OnDestroy, OnChanges {
     private fetchForm(): void {
         if (this.action) {
             this.isLoading = true;
-
             this.subscription = this.http.get<AjaxFormResponse>(this.action).subscribe({
                 next: (response) => this.responseHandler(response),
                 error: (response) => this.responseHandler(response),