+++ libs/datasource.inline/src/lib/datasource-inline.stories.ts
@@ -1,8 +1,14 @@
 import { Component, importProvidersFrom, Injector, Input } from '@angular/core';
+import { DatasourceModule, DatasourceService } from '@spryker/datasource';
 import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
 import { of } from 'rxjs';
-import { DatasourceModule, DatasourceService } from '@spryker/datasource';
 import { DatasourceInlineService } from './datasource-inline.service';
+
+const context = {
+    dependable1: 'data-1',
+    dependable2: 'data-2',
+    dependable3: 'data-3',
+};
 
 @Component({
     selector: 'spy-test',
@@ -15,6 +21,8 @@ import { DatasourceInlineService } from './datasource-inline.service';
 })
 class TestComponent {
     @Input() datasourceDataProp = '';
+    @Input() dependable = false;
+    @Input() contextKey = '';
 
     datasourceData = of('initial data');
 
@@ -24,10 +32,28 @@ class TestComponent {
     ) {}
 
     getData(): void {
-        this.datasourceData = this.datasourceService.resolve(this.injector, {
-            type: 'inline',
-            data: this.datasourceDataProp,
-        } as any);
+        this.datasourceData = this.datasourceService.resolve(
+            this.injector,
+            {
+                type: 'inline',
+                data: !this.dependable
+                    ? this.datasourceDataProp
+                    : {
+                          'data-1': 'data depends on context 1',
+                          'data-2': 'data depends on context 2',
+                          'data-3': 'data depends on context 3',
+                      },
+                ...(this.dependable
+                    ? {
+                          dependsOnContext: {
+                              contextKey: this.contextKey,
+                              default: 'default data',
+                          },
+                      }
+                    : {}),
+            },
+            this.dependable ? context : undefined,
+        );
     }
 }
 
@@ -47,14 +73,38 @@ export default {
             declarations: [TestComponent],
         }),
     ],
+    parameters: {
+        controls: {
+            include: ['contextKey', 'datasourceDataProp'],
+        },
+    },
     args: {
-        datasourceDataProp: 'new data',
+        datasourceDataProp: '',
+        contextKey: '',
     },
 } as Meta;
 
 export const primary = (args) => ({
     props: args,
     template: `
-    <spy-test [datasourceDataProp]="datasourceDataProp"></spy-test>
-  `,
+        <spy-test [datasourceDataProp]="datasourceDataProp"></spy-test>
+    `,
+});
+
+export const dependable = (args) => ({
+    props: {
+        ...args,
+        context: context,
+    },
+    template: `
+        use contextKey to set the context key depends on context
+        <br />
+        context: {{ context | json }}
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <spy-test [dependable]="true" [contextKey]="contextKey"></spy-test>
+    `,
 });