+++ libs/datasource.trigger.change/src/lib/change-datasource-trigger.stories.ts
@@ -1,5 +1,5 @@
 import { AfterViewInit, Component, importProvidersFrom, Injector, Input, OnInit, ViewChild } from '@angular/core';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
 import { MockHttpModule, setMockHttp } from '@spryker/internal-utils';
@@ -13,6 +13,7 @@ import {
 } from '@spryker/datasource.trigger';
 import { Observable, of, ReplaySubject } from 'rxjs';
 import { ChangeDatasourceTriggerService } from './change-datasource-trigger.service';
+import { provideHttpClient } from '@angular/common/http';
 
 const mockOptions = [
     {
@@ -85,7 +86,8 @@ export default {
         applicationConfig({
             providers: [
                 provideAnimations(),
-                importProvidersFrom(HttpClientTestingModule),
+                provideHttpClient(),
+                provideHttpClientTesting(),
                 importProvidersFrom(
                     DatasourceModule.withDatasources({
                         trigger: DatasourceTriggerService,