+++ libs/table.column.chip/src/lib/table-column-chip.component.stories.ts
@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { importProvidersFrom } from '@angular/core';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
@@ -9,6 +9,7 @@ import { ContextModule, DefaultContextSerializationModule } from '@spryker/utils
 
 import { TableColumnChipComponent } from './table-column-chip.component';
 import { TableColumnChipModule } from './table-column-chip.module';
+import { provideHttpClient } from '@angular/common/http';
 
 export default {
     title: 'TableColumnChipComponent',
@@ -56,7 +57,8 @@ export const withTable = (args) => ({
     props: args,
     applicationConfig: {
         providers: [
-            importProvidersFrom(HttpClientTestingModule),
+            provideHttpClient(),
+            provideHttpClientTesting(),
             importProvidersFrom(TableModule.forRoot()),
             importProvidersFrom(
                 TableModule.withColumnComponents({