+++ libs/table.feature.title/src/lib/table-title-feature.component.stories.ts
@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { importProvidersFrom } from '@angular/core';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
@@ -8,6 +8,7 @@ import { MockTableDatasourceConfig, MockTableDatasourceService, TableDataMockGen
 import { DefaultContextSerializationModule } from '@spryker/utils';
 
 import { TableTitleFeatureModule } from './table-title-feature.module';
+import { provideHttpClient } from '@angular/common/http';
 
 const tableDataGenerator: TableDataMockGenerator = (i) => ({
     col1: `col1 #${i}`,
@@ -21,7 +22,8 @@ export default {
         applicationConfig({
             providers: [
                 provideAnimations(),
-                importProvidersFrom(HttpClientTestingModule),
+                provideHttpClient(),
+                provideHttpClientTesting(),
                 importProvidersFrom(TableModule.forRoot()),
                 importProvidersFrom(
                     DatasourceModule.withDatasources({