+++ libs/tag/src/lib/tag/tag.component.spec.ts
@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { Component, DebugElement, NO_ERRORS_SCHEMA } from '@angular/core';
 import { By } from '@angular/platform-browser';
@@ -22,7 +22,7 @@ describe('TagComponent', () => {
     let fixture: ComponentFixture<TestComponent>;
     let labelElement: DebugElement;
 
-    beforeEach(async(() => {
+    beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
             declarations: [TagComponent, TestComponent],
             schemas: [NO_ERRORS_SCHEMA],