+++ libs/ajax-action/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2](http://172.31.0.22:9292/spryker-internal-ci/ui-components/compare/@spryker/ajax-action@2.0.1...@spryker/ajax-action@2.0.2) (2025-05-20)
+
+**Note:** Version bump only for package @spryker/ajax-action
+
+
+
+
+
+## [2.0.1](http://172.31.0.22:9292/spryker-internal-ci/ui-components/compare/@spryker/ajax-action@2.0.0...@spryker/ajax-action@2.0.1) (2024-12-11)
+
+**Note:** Version bump only for package @spryker/ajax-action
+
+
+
+
+
 # [2.0.0](https://github.com/spryker/ui-components/compare/@spryker/ajax-action@2.0.0-rc.1...@spryker/ajax-action@2.0.0) (2024-03-21)
 
 **Note:** Version bump only for package @spryker/ajax-action