+++ libs/tree-select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.10](http://172.31.0.22:9292/spryker-internal-ci/ui-components/compare/@spryker/tree-select@2.0.9...@spryker/tree-select@2.0.10) (2025-05-20)
+
+**Note:** Version bump only for package @spryker/tree-select
+
+
+
+
+
+## [2.0.9](http://172.31.0.22:9292/spryker-internal-ci/ui-components/compare/@spryker/tree-select@2.0.8...@spryker/tree-select@2.0.9) (2024-12-11)
+
+**Note:** Version bump only for package @spryker/tree-select
+
+
+
+
+
 ## [2.0.8](http://172.31.33.130:9292/spryker-internal-ci/ui-components/compare/@spryker/tree-select@2.0.7...@spryker/tree-select@2.0.8) (2024-09-25)
 
 **Note:** Version bump only for package @spryker/tree-select